Refactor the ZUIAvatar component so it's possible to use just the background.
