+++ src/zui/components/ZUIAvatar/index.tsx
@@ -1,9 +1,8 @@
-import { create } from 'random-seed';
-import { FC, useMemo } from 'react';
+import { FC } from 'react';
 import { useTheme } from '@mui/material';
 
-import { funSwatches } from 'zui/theme/palette';
 import { ZUISize } from '../types';
+import AvatarBackground from './AvatarBackground';
 
 export interface ZUIAvatarProps {
   /**
@@ -54,106 +53,23 @@ const ZUIAvatar: FC<ZUIAvatarProps> = ({
     letterSpacing = 0.14;
   }
 
-  const seededRand = create(id.toString());
-  const rand = () => seededRand(1000000) / 1000000;
   const initials = firstName.slice(0, 1) + lastName.slice(0, 1);
 
-  const colors = useMemo(() => {
-    return Object.keys(funSwatches).reduce((acc, swatch) => {
-      acc.push(funSwatches[swatch].light.color);
-      acc.push(funSwatches[swatch].medium.color);
-
-      return acc;
-    }, [] as string[]);
-  }, [funSwatches]);
-
-  const getColor = (): string => {
-    const index = seededRand(colors.length);
-    const color = colors[index];
-    return color;
-  };
-
-  const uniqueName = (name: string) => {
-    return name + id;
-  };
-
   return (
-    <svg
-      height={avatarSize}
-      version="1.1"
-      width={avatarSize}
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <defs>
-        <linearGradient
-          gradientTransform={'rotate(' + rand() * 360 + ')'}
-          id={uniqueName('gradient')}
-        >
-          <stop offset="5%" stopColor={getColor()} />
-          <stop offset="95%" stopColor={getColor()} />
-        </linearGradient>
-
-        <filter id={uniqueName('blurMe')}>
-          <feGaussianBlur
-            edgeMode="wrap"
-            in="SourceGraphic"
-            stdDeviation="10"
-          />
-        </filter>
-
-        <clipPath id={uniqueName(size + 'circleClip')}>
-          <circle cx={avatarSize / 2} cy={avatarSize / 2} r={avatarSize / 2} />
-        </clipPath>
-        <clipPath id={uniqueName(size + 'squareClip')}>
-          <rect height={avatarSize} rx="4" ry="4" width={avatarSize} />
-        </clipPath>
-      </defs>
-
-      <g
-        clipPath={
-          variant == 'circular'
-            ? `url(#${uniqueName(size + 'circleClip')})`
-            : `url(#${uniqueName(size + 'squareClip')})`
-        }
-        id={'content' + id}
+    <AvatarBackground seed={id.toString()} size={avatarSize} variant={variant}>
+      <text
+        dominantBaseline="central"
+        fontFamily={theme.typography.fontFamily}
+        fontSize={fontSize}
+        fontWeight={600}
+        letterSpacing={letterSpacing}
+        textAnchor="middle"
+        x={avatarSize / 2}
+        y={avatarSize / 2}
       >
-        <rect
-          fill={`url(#${uniqueName('gradient')})`}
-          height={avatarSize}
-          width={avatarSize}
-        />
-        <g filter={`url(#${uniqueName('blurMe')})`}>
-          <rect
-            fill={getColor()}
-            height={avatarSize}
-            width={avatarSize}
-            x={-seededRand(avatarSize)}
-            y={-seededRand(avatarSize) / 2}
-          />
-          <line
-            stroke={getColor()}
-            strokeWidth={15}
-            x1={avatarSize * 1.5}
-            x2={avatarSize / 2}
-            y1={avatarSize * 1.5}
-            y2={avatarSize / 3}
-          />
-        </g>
-
-        <text
-          dominantBaseline="central"
-          fontFamily={theme.typography.fontFamily}
-          fontSize={fontSize}
-          fontWeight={600}
-          letterSpacing={letterSpacing}
-          textAnchor="middle"
-          x={avatarSize / 2}
-          y={avatarSize / 2}
-        >
-          {initials.toUpperCase()}
-        </text>
-      </g>
-    </svg>
+        {initials.toUpperCase()}
+      </text>
+    </AvatarBackground>
   );
 };
